// types.ts
import { Db, MongoClient } from 'mongodb';
import { NextApiRequest, NextApiResponse } from 'next';

export interface PostFrontMatter {
  title: string;
  summary: string;
  publishedOn: string;
}

export interface Post {
  source: string;
  frontMatter: PostFrontMatter;
}

export interface UserSession {
  id: string;
  image: string;
  email: string;
  name: string;
}

export interface Request extends NextApiRequest {
  db: Db;
  dbClient: MongoClient;
  user: { email: string; id: string };
}

export type Response = NextApiResponse;